Refactor querylayer and cleanup Postgres migration
*Split monolithic clarity_queries.py (1341 lines) into domain-specific
modules under integrations/queries/ (projects, samples, sequencing,
storage).
*Push column renaming into the query layer using display-name
dict keys, removing all df.rename() calls from data_utils.py.
*Extract DatasetCache class in app.py for per-session reactive caching, and
extract storage tool into features/storage.py.
* Remove dashboard refresh button (detail toolbar refresh is sufficient).
* Fix stale column name reference in sequencing feature (Cluster Density).
