+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "shinylims-uv"
-version = "1.2.1"
+version = "1.3.0"
 authors = [
   { name="Magnus Leithaug", email="magnus.leithaug@vetinst.no" },
 ]
