+++ src/shinylims/integrations/data_utils.py
@@ -6,7 +6,7 @@ This module provides functions to fetch and transform live Clarity Postgres
 data for projects, samples, sequencing runs, and storage containers.
 Historical samples are still read from a SQLite pin.
 
-Also includes functions to transform LIMS IDs and comments into HTML
+Also includes functions to transform LIMS IDs and comments into HTML.
 '''
 
 ####################
@@ -25,7 +25,7 @@ import pandas as pd
 import tomli
 
 from shinylims.integrations.clarity_pg import create_session
-from shinylims.integrations.clarity_queries import (
+from shinylims.integrations.queries import (
     build_project_rows,
     build_sample_rows,
     build_sequencing_run_rows,
@@ -123,33 +123,7 @@ def _log_pg_fetch_timing(scope: str, started_at: float, **metrics: object) -> No
 
 
 def _format_sequencing_dataframe(df: pd.DataFrame) -> pd.DataFrame:
-    """Normalize sequencing rows to the app-facing dataframe shape."""
-    df = df.rename(columns={
-        'run_id': 'Run ID',
-        'instrument': 'Instrument',
-        'run_number': 'Run Number',
-        'seq_date': 'Seq Date',
-        'operator': 'Operator',
-        'species': 'Species',
-        'experiment_name': 'Experiment Name',
-        'casette_type': 'Casette Type',
-        'read_length': 'Read Length',
-        'index_cycles': 'Index Cycles',
-        'sample_count': 'Sample Count',
-        'loading_pm': 'Loading pM',
-        'diluted_denatured_ul': 'Diluted Denatured (uL)',
-        'avg_fragment_size': 'Avg Fragment Size',
-        'combined_pool': 'Combined Pool',
-        'phix_loaded_percent': 'Phix Loaded (%)',
-        'phix_aligned_percent': 'Phix Aligned (%)',
-        'cluster_density': 'Cluster Density',
-        'yield_total' : 'Yield Total',
-        'qv30_r1': 'QV30 R1',
-        'qv30_r2':  'QV30 R2',
-        'pf_reads': 'PF Reads',
-        'comment' : 'Comment'
-    })
-
+    """Apply HTML transforms and sanitization to a sequencing DataFrame."""
     if 'Seq Date' in df.columns:
         df['Seq Date'] = pd.to_datetime(df['Seq Date'], errors='coerce')
 
@@ -167,36 +141,7 @@ def _format_sequencing_dataframe(df: pd.DataFrame) -> pd.DataFrame:
 
 
 def _format_samples_dataframe(df: pd.DataFrame, *, meta_created: str) -> tuple[pd.DataFrame, str]:
-    """Normalize sample rows to the app-facing dataframe shape."""
-    if "storage_box_formatted" not in df.columns:
-        df["storage_box_formatted"] = df.get("storage_box", "")
-
-    if 'storage_box' in df.columns:
-        df = df.drop('storage_box', axis=1)
-
-    df = df.rename(columns={
-        'limsid': 'LIMS ID',
-        'project_limsid': 'Project LIMS ID',
-        'received_date': 'Received Date',
-        'species_name': 'Species',
-        'name': 'Sample Name',
-        'project_name': 'Project Name',
-        'project_account': 'Project Account',
-        'experiment_name': 'Experiment Name',
-        'invoice_id': 'Invoice ID',
-        'extraction_number': 'Extraction Number',
-        'concentration_absorbance': 'Absorbance',
-        'a260_280_ratio': 'A260/280 ratio',
-        'a260_230_ratio': 'A260/230 ratio',
-        'concentration_fluorescence': 'Fluorescence',
-        'storage_box_formatted': 'Storage Box',
-        'storage_well': 'Storage Well',
-        'billing_description': 'Billing Description',
-        'reagent_label': 'Reagent Label',
-        'increased_pooling': 'Increased Pooling (%)',
-        'nird_filename': 'NIRD Filename',
-    })
-
+    """Apply HTML transforms, sanitization, and column ordering to a samples DataFrame."""
     if 'Received Date' in df.columns:
         df['Received Date'] = pd.to_datetime(df['Received Date'], errors='coerce')
 
@@ -215,6 +160,7 @@ def _format_samples_dataframe(df: pd.DataFrame, *, meta_created: str) -> tuple[p
     } | ({'Storage Box'} if 'Storage Box' in df.columns else set())
     df = sanitize_dataframe_strings(df, skip_columns=html_columns)
 
+    # Place Storage Box immediately before Storage Well
     cols = df.columns.tolist()
     if 'Storage Box' in cols:
         cols.remove('Storage Box')
@@ -227,7 +173,7 @@ def _format_samples_dataframe(df: pd.DataFrame, *, meta_created: str) -> tuple[p
 
 
 def _fetch_projects_data_from_postgres() -> tuple[pd.DataFrame, str]:
-    """Fetch project rows directly from Clarity Postgres using the prototype query layer."""
+    """Fetch project rows directly from Clarity Postgres."""
     started_at = time.perf_counter()
     with create_session() as session:
         rows = build_project_rows(session)
@@ -235,22 +181,22 @@ def _fetch_projects_data_from_postgres() -> tuple[pd.DataFrame, str]:
     df = pd.DataFrame(rows)
     if df.empty:
         df = pd.DataFrame(columns=[
-            "project_limsid",
-            "open_date",
-            "status",
-            "project_name",
-            "sample_count",
-            "species",
-            "submitter",
-            "submitting_lab",
-            "comment",
+            "Project LIMS ID",
+            "Open Date",
+            "Status",
+            "Project Name",
+            "Samples",
+            "Species",
+            "Submitter",
+            "Submitting Lab",
+            "Comment",
         ])
     _log_pg_fetch_timing("fetch_projects_data_from_postgres", started_at, row_count=len(df))
     return df, datetime.datetime.now().isoformat()
 
 
 def _fetch_sequencing_data_from_postgres() -> tuple[pd.DataFrame, str]:
-    """Fetch sequencing rows directly from Clarity Postgres using the prototype query layer."""
+    """Fetch sequencing rows directly from Clarity Postgres."""
     sequencing_type_ids = _get_clarity_pg_sequencing_type_ids()
     if not sequencing_type_ids:
         raise RuntimeError(
@@ -265,29 +211,29 @@ def _fetch_sequencing_data_from_postgres() -> tuple[pd.DataFrame, str]:
     if df.empty:
         df = pd.DataFrame(columns=[
             "seq_limsid",
-            "run_id",
-            "instrument",
-            "run_number",
-            "seq_date",
-            "operator",
-            "species",
-            "experiment_name",
-            "casette_type",
-            "read_length",
-            "index_cycles",
-            "sample_count",
-            "loading_pm",
-            "diluted_denatured_ul",
-            "avg_fragment_size",
-            "combined_pool",
-            "phix_loaded_percent",
-            "phix_aligned_percent",
-            "cluster_density",
-            "yield_total",
-            "qv30_r1",
-            "qv30_r2",
-            "pf_reads",
-            "comment",
+            "Run ID",
+            "Instrument",
+            "Run Number",
+            "Seq Date",
+            "Operator",
+            "Species",
+            "Experiment Name",
+            "Casette Type",
+            "Read Length",
+            "Index Cycles",
+            "Sample Count",
+            "Loading pM",
+            "Diluted Denatured (uL)",
+            "Avg Fragment Size",
+            "Combined Pool",
+            "Phix Loaded (%)",
+            "Phix Aligned (%)",
+            "Cluster Density",
+            "Yield Total",
+            "QV30 R1",
+            "QV30 R2",
+            "PF Reads",
+            "Comment",
         ])
     _log_pg_fetch_timing(
         "fetch_sequencing_data_from_postgres",
@@ -299,7 +245,7 @@ def _fetch_sequencing_data_from_postgres() -> tuple[pd.DataFrame, str]:
 
 
 def _fetch_samples_data_from_postgres() -> tuple[pd.DataFrame, str]:
-    """Fetch sample rows directly from Clarity Postgres using the prototype query layer."""
+    """Fetch sample rows directly from Clarity Postgres."""
     started_at = time.perf_counter()
     with create_session() as session:
         rows = build_sample_rows(session, open_projects_only=False)
@@ -307,38 +253,37 @@ def _fetch_samples_data_from_postgres() -> tuple[pd.DataFrame, str]:
     df = pd.DataFrame(rows)
     if df.empty:
         df = pd.DataFrame(columns=[
-            "limsid",
-            "project_limsid",
-            "received_date",
+            "LIMS ID",
+            "Project LIMS ID",
+            "Received Date",
             "progress",
-            "species_name",
-            "name",
-            "project_name",
+            "Species",
+            "Sample Name",
+            "Project Name",
             "submitter",
             "submitting_lab",
-            "project_account",
-            "experiment_name",
-            "extraction_number",
-            "concentration_absorbance",
-            "a260_280_ratio",
-            "a260_230_ratio",
-            "concentration_fluorescence",
-            "storage_box",
-            "storage_box_formatted",
-            "storage_well",
-            "invoice_id",
+            "Project Account",
+            "Experiment Name",
+            "Extraction Number",
+            "Absorbance",
+            "A260/280 ratio",
+            "A260/230 ratio",
+            "Fluorescence",
+            "Storage Box",
+            "Storage Well",
+            "Invoice ID",
             "sample_type",
             "gram_stain",
-            "nird_filename",
-            "billing_description",
+            "NIRD Filename",
+            "Billing Description",
             "price",
             "nd_limsid",
             "qubit_limsid",
             "prep_limsid",
             "seq_limsid",
             "billed_limsid",
-            "increased_pooling",
-            "reagent_label",
+            "Increased Pooling (%)",
+            "Reagent Label",
         ])
     _log_pg_fetch_timing("fetch_samples_data_from_postgres", started_at, row_count=len(df))
     return df, datetime.datetime.now().isoformat()
@@ -353,16 +298,14 @@ def fetch_storage_containers_data() -> pd.DataFrame:
     if df.empty:
         df = pd.DataFrame(
             columns=[
-                "container_name",
-                "status",
-                "created_date",
-                "last_modified_date",
+                "Box Name",
+                "Status",
+                "Created Date",
+                "Last Modified",
             ]
         )
     _log_pg_fetch_timing("fetch_storage_containers_data_from_postgres", started_at, row_count=len(df))
     return df
-
-
 
 
 ####################
@@ -373,34 +316,18 @@ def fetch_storage_containers_data() -> pd.DataFrame:
 def fetch_projects_data():
     """Fetch project data directly from Clarity Postgres."""
     df, meta_created = _fetch_projects_data_from_postgres()
-    
-    # Rename columns to match what the app expects
-    df = df.rename(columns={
-        'open_date': 'Open Date',
-        'project_name': 'Project Name',
-        'sample_count': 'Samples',
-        'submitting_lab': 'Submitting Lab',
-        'comment': 'Comment',
-        'status': 'Status',
-        'submitter': 'Submitter',
-        'species': 'Species',
-        'project_limsid': 'Project LIMS ID',
-    })
 
-    # Apply transformations
     if 'Open Date' in df.columns:
         df['Open Date'] = pd.to_datetime(df['Open Date'], errors='coerce')
 
-    # Replace NaN values
     df = df.replace(np.nan, '', regex=True)
 
-    # Transform comments
     comment_columns = [col for col in df.columns if 'comment' in col.lower()]
     for col in comment_columns:
         df[col] = df[col].apply(transform_comments_to_html)
 
     df = sanitize_dataframe_strings(df, skip_columns=set(comment_columns))
-    
+
     return df, meta_created
 
 
@@ -411,13 +338,10 @@ def fetch_all_samples_data():
     return _format_samples_dataframe(df, meta_created=meta_created)
 
 
-
-
-
 def fetch_historical_samples_data():
     """
     Fetch historical samples data from SQLite database.
-    Returns: tuple of (DataFrame, creation_date)
+    Returns: DataFrame
     """
 
     df = query_to_dataframe("SELECT * FROM samples_historical")
@@ -445,20 +369,16 @@ def fetch_historical_samples_data():
         'increased_pooling': 'Increased Pooling (%)',
         'nird_filename': 'NIRD Filename',
     })
-    
-    # Apply transformations
+
     if 'Received Date' in df.columns:
         df['Received Date'] = pd.to_datetime(df['Received Date'], errors='coerce')
-    
-    # Replace NaN values
+
     df = df.replace(np.nan, '', regex=True)
-    
-    # Transform LIMS IDs to HTML links
+
     for col in ['seq_limsid', 'nd_limsid', 'qubit_limsid', 'prep_limsid']:
         if col in df.columns:
             df[col] = df[col].apply(transform_to_html)
-    
-    # Transform comments
+
     comment_columns = [col for col in df.columns if 'comment' in col.lower()]
     for col in comment_columns:
         df[col] = df[col].apply(transform_comments_to_html)
@@ -482,29 +402,27 @@ def get_app_version():
     try:
         # Look in the same directory as app.py
         toml_path = Path(__file__).parent / "pyproject.toml"
-        
+
         # If not found there, try one level up
         if not toml_path.exists():
             toml_path = Path(__file__).resolve().parent.parent / "pyproject.toml"
-            
+
         # If still not found, try the current working directory
         if not toml_path.exists():
             toml_path = Path.cwd() / "pyproject.toml"
-        
+
         if toml_path.exists():
             with open(toml_path, "rb") as f:
                 data = tomli.load(f)
-                # Try different possible locations of version information
                 if "tool" in data and "poetry" in data["tool"]:
                     return data["tool"]["poetry"].get("version", "Unknown")
                 elif "project" in data:
                     return data["project"].get("version", "Unknown")
                 else:
-                    # Direct search for version if not in expected locations
                     for section in data:
                         if "version" in data[section]:
                             return data[section]["version"]
-        
+
         return "Unknown"
-    except Exception as e:
+    except Exception:
         return "Unknown"
